- Move config to conf page - Add an URL rewriting system to have beautiful URL - Fix some CSS - Fix some JS issues regarding the URL rewrite (Ajax request)

+++ system/Widget/WidgetWrapper.php
@@ -67,8 +67,10 @@ class WidgetWrapper
 
     static function destroyInstance()
     {
-        self::$instance->destroy();
-        self::$instance = null;
+        if(isset(self::$instance)) {
+            self::$instance->destroy();
+            self::$instance = null;
+        }
     }
 
     /**
